Update formatAmount in TxSummary

+++ app/src/components/TxSummary.tsx
@@ -59,7 +59,7 @@ const TxSummary: FC<TxSummaryProps> = ({
                   </div>
                   {fees.inDollars > 0 && (
                     <div className="text-right text-turtle-level4">
-                      ${formatAmount(fees.inDollars, 'Long')}
+                      ${formatAmount(fees.inDollars, 'Short')}
                     </div>
                   )}
                 </div>
